move growing protocol tests to own file
